+++ Mammoth/TSE/CDesignType.cpp
@@ -3669,7 +3669,7 @@ ALERROR CEffectCreatorRef::Bind (SDesignLoadCtx &Ctx)
 			}
 		catch (...)
 			{
-			::kernelDebugLogPattern("Crash deleting singleton: %08x. UNID = %08x", (DWORD)m_pSingleton, m_dwUNID);
+			::kernelDebugLogPattern("Crash deleting singleton: %08x. UNID = %08x", (DWORD)(uintptr_t)m_pSingleton, m_dwUNID);
 			throw;
 			}
