+++ Mammoth/TSE/CMissile.cpp
@@ -460,7 +460,7 @@ CString CMissile::DebugCrashInfo (void) const
 		}
 	catch (...)
 		{
-		sResult.Append(strPatternSubst(CONSTLIT("m_pExhaust: %x [invalid]\r\n"), (DWORD)m_pExhaust));
+		sResult.Append(strPatternSubst(CONSTLIT("m_pExhaust: %x [invalid]\r\n"), (DWORD)(uintptr_t)m_pExhaust));
 		}
 
 	//	m_pEnhancements
@@ -474,7 +474,7 @@ CString CMissile::DebugCrashInfo (void) const
 		}
 	catch (...)
 		{
-		sResult.Append(strPatternSubst(CONSTLIT("m_pEnhancements: %x [invalid]\r\n"), (DWORD)(CItemEnhancementStack *)m_pEnhancements));
+		sResult.Append(strPatternSubst(CONSTLIT("m_pEnhancements: %x [invalid]\r\n"), (DWORD)(uintptr_t)(CItemEnhancementStack *)m_pEnhancements));
 		}
 
 	//	m_pPainter
@@ -488,7 +488,7 @@ CString CMissile::DebugCrashInfo (void) const
 		}
 	catch (...)
 		{
-		sResult.Append(strPatternSubst(CONSTLIT("m_pPainter: %x [invalid]\r\n"), (DWORD)m_pPainter));
+		sResult.Append(strPatternSubst(CONSTLIT("m_pPainter: %x [invalid]\r\n"), (DWORD)(uintptr_t)m_pPainter));
 		}
 
 	//	Objects
