+++ Mammoth/TSE/CRadiusDamage.cpp
@@ -216,7 +216,7 @@ CString CRadiusDamage::DebugCrashInfo (void) const
 		}
 	catch (...)
 		{
-		sResult.Append(strPatternSubst(CONSTLIT("m_pEnhancements: %x [invalid]\r\n"), (DWORD)(CItemEnhancementStack *)m_pEnhancements));
+		sResult.Append(strPatternSubst(CONSTLIT("m_pEnhancements: %x [invalid]\r\n"), (DWORD)(uintptr_t)(CItemEnhancementStack *)m_pEnhancements));
 		}
 
 	//	m_pPainter
@@ -230,7 +230,7 @@ CString CRadiusDamage::DebugCrashInfo (void) const
 		}
 	catch (...)
 		{
-		sResult.Append(strPatternSubst(CONSTLIT("m_pPainter: %x [invalid]\r\n"), (DWORD)m_pPainter));
+		sResult.Append(strPatternSubst(CONSTLIT("m_pPainter: %x [invalid]\r\n"), (DWORD)(uintptr_t)m_pPainter));
 		}
 
 	return sResult;
