+++ Mammoth/TSE/CSpaceObject.cpp
@@ -1545,15 +1545,15 @@ CString CSpaceObject::DebugDescribe (CSpaceObject *pObj)
 		if (pObj == NULL)
 			return CONSTLIT("none");
 		else if (pObj->IsDestroyed())
-			return strPatternSubst(CONSTLIT("%x %s (%s) [destroyed]"), (DWORD)pObj, pObj->GetNounPhrase(), pObj->GetObjClassName());
+			return strPatternSubst(CONSTLIT("%x %s (%s) [destroyed]"), (DWORD)(uintptr_t)pObj, pObj->GetNounPhrase(), pObj->GetObjClassName());
 		else
-			return strPatternSubst(CONSTLIT("%x %s (%s)"), (DWORD)pObj, pObj->GetNounPhrase(), pObj->GetObjClassName());
+			return strPatternSubst(CONSTLIT("%x %s (%s)"), (DWORD)(uintptr_t)pObj, pObj->GetNounPhrase(), pObj->GetObjClassName());
 		}
 	catch (...)
 		{
 		}
 
-	return strPatternSubst(CONSTLIT("%x [invalid]"), (DWORD)pObj);
+	return strPatternSubst(CONSTLIT("%x [invalid]"), (DWORD)(uintptr_t)pObj);
 	}
 
 CString CSpaceObject::DebugLoadError (SLoadCtx &Ctx)
